% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinStats.R
\name{basinStats}
\alias{basinStats}
\title{Gets statistics of .asc file of a basin}
\usage{
basinStats(asc_file)
}
\arguments{
\item{asc_file}{Required. A text \code{.asc} file describing a basin, as used by \code{WDPM}.}
}
\value{
If successful, returns a vector containing the following values
\describe{
 \item{num_rows}{Number of rows in the file}
 \item{num_cols}{Number of cols in the file}
 \item{cell_size}{Width of each cell. Typically in metres, could be in degrees}
 \item{outside_cells}{Number of cells outside the basin}
 \item{inside_cells}{Number of cells inside the basin}
 \item{inside_area}{Basin area in units of \code{cell_size} squared}
}
}
\description{
Returns the statistics describing a basin contained in an ArcGIS \code{.asc} file,
either a DEM or a water layer. The points outside the basin must be coded as missing.
}
\examples{
\dontrun{
stats <- basinStats("StDenis.asc")
}
}
