% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributingAreaHysteresisPlot.R
\name{contributingAreaHysteresisPlot}
\alias{contributingAreaHysteresisPlot}
\title{Plots hysteresis curves of connected/contributing area against volume}
\usage{
contributingAreaHysteresisPlot(curve1, curve2, extend = TRUE,
  normalize = TRUE, outfile = "", N = 1)
}
\arguments{
\item{curve1}{Required. Data frame containing the filling curve.}

\item{curve2}{Required. Data frame containing the emptying curve.}

\item{extend}{Optional. Should curves extend to origin? Default is \code{TRUE}.}

\item{normalize}{Optional. Should areas and volumes be divided by their maximum values? Default is \code{TRUE}.}

\item{outfile}{Optional. File name for output of line segments. If not specified (the default), the data are not ouput.}

\item{N}{Optional. Plot arrows every Nth point. Default is 1.}
}
\value{
Returns a ggplot2 object of the area vs volume. The line segments are written to a file (if specified).
}
\description{
Plots hysteresis curves of connected/contributing area against volume
}
\examples{
\dontrun{
p <- contributingAreaHysteresisPlot(curve1, curve2)}

}
