% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleContribFracPlot.R
\name{doubleContribFracPlot}
\alias{doubleContribFracPlot}
\title{Plots 2 sets of contributing fraction hysteresis curves}
\usage{
doubleContribFracPlot(curveFile1, name1, curveFile2, name2)
}
\arguments{
\item{curveFile1}{Required. File containing first set of curves.}

\item{name1}{Required. Name of first set of curves.}

\item{curveFile2}{Required. File containing second set of curves.}

\item{name2}{Required. Name of second set of curves.}
}
\value{
Returns \pkg{ggplot2} object of both sets of curves
}
\description{
Plots 2 sets of contributing fraction hysteresis curves
}
\examples{
\dontrun{
p <- doubleContribFracPlot('Run1_loop1.csv', 'PCM', 'WDPM_loop1.csv', 'WDPM')}
}
