% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClosestPointToOutlet.R
\name{findClosestPointToOutlet}
\alias{findClosestPointToOutlet}
\title{Finds minumum distance to outet from pond}
\usage{
findClosestPointToOutlet(water_file, outletXY)
}
\arguments{
\item{water_file}{Required. WDPM output.}

\item{outletXY}{Required. Vector containing outlet X and Y coordinates. Can be
the output of the command \code{findOutlet}.}
}
\value{
Returns the minimum distance from the the pond divide to the outlet.
}
\description{
Finds minumum distance to outet from pond
}
\examples{
\dontrun{
outlet <- findOutlet("basin5.asc")
min_distance <- findClosestPointToOutlet("100_0_0_0_d.asc", outlet$pourXY)
}
}
\seealso{
\code{\link{findOutlet}}
}
\author{
Kevin Shook
}
