% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFarthest.R
\name{findFarthest}
\alias{findFarthest}
\title{Finds location of farthest point from outlet}
\usage{
findFarthest(DEMfile)
}
\arguments{
\item{DEMfile}{Required. ArcGIS \code{.asc} file of a basin DEM. Must have non-basin points indicated as missing.}
}
\value{
Returns a list containing the XY coordinates and the distance of the point farthest from the outlet.
}
\description{
Finds location of farthest point from outlet
}
\examples{
\dontrun{
farthest <- findFarthest("basin5.asc")
}
}
\seealso{
\code{\link{findOutlet}}
}
