% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLargestPondCentroid.R
\name{findLargestPondCentroid}
\alias{findLargestPondCentroid}
\title{Finds centroid of largest water area}
\usage{
findLargestPondCentroid(water_file, threshold = 0.001)
}
\arguments{
\item{water_file}{Required. WDPM output.}

\item{threshold}{Optional. Threshold for delineating water. Default is 0.001 m.}
}
\value{

}
\description{
Finds centroid of largest water area
}
\examples{
\dontrun{
centroid <- findLargestPondCentroid("100_0_0_0_d.asc")
}
}
