% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOutlet.R
\name{findOutlet}
\alias{findOutlet}
\title{Finds lowest elevation in a DEM file, and the basin pour point}
\usage{
findOutlet(DEMfile)
}
\arguments{
\item{DEMfile}{Required. ArcGIS \code{.asc} file of a basin DEM. Must have non-basin points indicated as missing.}
}
\value{
Returns a list containing the mininum elevation within
}
\description{
Finds lowest elevation in a DEM file, and the basin pour point
}
\examples{
\dontrun{
outlet <- findOutlet('basin5.asc')
}
}
\seealso{
\code{\link{findFarthest}} \code{\link{findClosestPointToOutlet}}
the basin, its location, the pour point elevation and its location.
}
