% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDegree3.R
\name{fitdegree3}
\alias{fitdegree3}
\title{Fits 3rd order polynomial to Korcak data}
\usage{
fitdegree3(korcakdata)
}
\arguments{
\item{korcakdata}{Required. Data frame produced by function \code{Korcak}.}
}
\value{
Returns coefficients of second-order polynomial (intercept, slope quadratic, r2).
}
\description{
Fits 3rd order polynomial to Korcak data
}
\examples{
\dontrun{
coeffs <- fitdegree3(korcakvals)}
}
\seealso{
\code{\link{korcak}}
}
\author{
Kevin Shook
}
