% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracBelowPour.R
\name{fracBelowPour}
\alias{fracBelowPour}
\title{Finds fraction of a basin below the pour point}
\usage{
fracBelowPour(DEMfile, fractionOnly = TRUE)
}
\arguments{
\item{DEMfile}{Required. ArcGIS \code{.asc} file of a basin DEM. Must have non-basin points indicated as missing.}

\item{fractionOnly}{Optional. If \code{TRUE} (the default) then only the fraction below the pour point is returned. If \code{FALSE}, then other statistics, and the raster below the pour point are returned}
}
\value{
Returns either the fraction of the basin below the pour point or a list containing the fraction of the basin area below the pour point and the rasters of the below-outlet dem and of the divide.
}
\description{
Finds fraction of a basin below the pour point
}
\examples{
\dontrun{
fracBelowPour('basin5.asc')
}
}
