% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWaterPatches.R
\name{getWaterPatches}
\alias{getWaterPatches}
\title{Gets water patches from an ArcGIS .asc file}
\usage{
getWaterPatches(infile, threshold = 0.001)
}
\arguments{
\item{infile}{Required. ArcGIS .asc file of water depths, as output by WDPM.}

\item{threshold}{Optional. Threshold (in m) to determine the existence of water. The default value is 0.001 m, i.e. 1 mm.}
}
\value{
Returns an SDMTools object containing the numbered water patches.
}
\description{
The purpose is to find water patches in an .asc file. The patch numbers can be used to determine the drainage basin areas.
}
\examples{
\dontrun{
patches <- getWaterPatches('10_0_0_0_d.asc')}
}
