% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotBasinAreaWidth.R
\name{ggplotBasinAreaWidth}
\alias{ggplotBasinAreaWidth}
\title{Create ggplot2 of basin width function}
\usage{
ggplotBasinAreaWidth(destinations, areas)
}
\arguments{
\item{destinations}{Required. Vector containing pond drainage destinations as used by \code{PCM}.}

\item{areas}{Required. Vector containing pond areas.}
}
\value{
Returns ggplot2 object of basin width function.
}
\description{
Create ggplot2 of basin width function
}
\examples{
\dontrun{p <- ggplotBasinWidth(dests, areas)}
}
