% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotBasinWidth.R
\name{ggplotBasinWidth}
\alias{ggplotBasinWidth}
\title{Create ggplot2 of basin width function}
\usage{
ggplotBasinWidth(destinations, counts = TRUE)
}
\arguments{
\item{destinations}{Required. Vector containing pond drainage destinations as used by \code{PCM}.}

\item{counts}{Optional. If \code{TRUE}, the default, plots number of ponds as y-axis. Otherwise plots the density.}
}
\value{
Returns ggplot2 object of basin width function.
}
\description{
Create ggplot2 of basin width function
}
\examples{
\dontrun{p <- ggplotBasinWidth(dests)}
}
