% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPCMPonds.R
\name{graphPCMPonds}
\alias{graphPCMPonds}
\title{Creates graph of PCM ponds}
\usage{
graphPCMPonds(dest, numPonds = 0, disConnected = "", size = 0,
  maxSize = 20, removeNodeZero = TRUE, labels = FALSE,
  arrowSize = 0, simplify = TRUE, vertexColours = "",
  edgeColours = "", outletRed = TRUE, lineWidth = 1)
}
\arguments{
\item{dest}{Required. Vector of drainage destinations.}

\item{numPonds}{Optional. Number of ponds to be plotted. If omitted, all ponds will be used.}

\item{disConnected}{Optional. Vector of link numbers to be deleted.}

\item{size}{Optional. Either scalar of common pond size or vector of all pond areas.}

\item{maxSize}{Optional. Used to scale vector of pond areas.}

\item{removeNodeZero}{Optional. Should pond zero be plotted? Default is \code{FALSE}.}

\item{labels}{Optional. Should ponds labels be plotted? Default is \code{FALSE}.}

\item{arrowSize}{Optional. Size of arrow heads. Default is \code{0} (no arrow).}

\item{simplify}{Optional. Should plot be simplified? Default is \code{TRUE}.}

\item{vertexColours}{Optional. Can be a single named colour for all vertices or a vector of names colours for individual ponds.}

\item{edgeColours}{Optional. Can be a single named colour for all edges or a vector of names colours for individual edges.}

\item{outletRed}{Optional. If \code{TRUE} (the default) then the outlet pond is coloured red.}

\item{lineWidth}{Optional. Sets line width for edges. Default is 1.}
}
\value{
Returns an \pkg{igraph} object of the PCM model.
}
\description{
Creates graph of PCM ponds
}
\examples{
\dontrun{
g <- graphPCMPonds(pcmModel)
plot(g)}
}
