% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysteresis.R
\name{hysteresis}
\alias{hysteresis}
\title{Calculates degree of hysteresis between 2 curves}
\usage{
hysteresis(curve1, curve2, extend = TRUE)
}
\arguments{
\item{curve1}{Required. Data frame of (x,y) values for first curve.}

\item{curve2}{Required. Data frame of (x,y) values for second curve.}

\item{extend}{Optional. Should curves be forced through the origin? Also extends the lower curve to include the maximum point. Default is \code{TRUE}.}
}
\value{
Returns the fractional hysteresis between the curves, i.e. the area between the curves as a fraction of the area below the upper curve.
}
\description{
Calculates fractional hysteresis as difference in areas under curves. Uses trapeziodal integration. Can force curves through origin, and adds max value to lower curve.
}
\examples{
\dontrun{
hysteresis(curve1, curve2)
}
}
\seealso{
\code{\link{trapint}}
}
\author{
Kevin Shook
}
