% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/korcak.R
\name{korcak}
\alias{korcak}
\title{Calculates non-exceedance probabilities}
\usage{
korcak(sizes, useLargest = FALSE, minval = 0)
}
\arguments{
\item{sizes}{Required. A vector of values.}

\item{useLargest}{Optional. If \code{TRUE} then the largest value will be used, i.e. its non-exceedence probability willl be > 0. Default is \code{FALSE}.}

\item{minval}{Optional. Minimum value to use. Default is \code{0}.}
}
\value{
Returns the size values and their non-exceedance probabilities
}
\description{
Calculates non-exceedance probabilities
}
\examples{
a <- korcak(runif(100))
}
