% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorcakPlot1.R
\name{KorcakPlot1}
\alias{KorcakPlot1}
\title{Korcak plot of 1 data set of water areas}
\usage{
KorcakPlot1(
  areas,
  minval = 100,
  addLine = "none",
  lineColour = "red",
  useLargest = FALSE
)
}
\arguments{
\item{areas}{Required. Vector of water areas.}

\item{minval}{Optional. Minimum water area to be plotted. Default is 100 m2.}

\item{addLine}{Optional. Default is \option{none}. Other options are \option{linear}, \option{quadratic}, \option{gpd} and
\option{pareto2}.}

\item{lineColour}{Optional. Colour to be used for the fitted line. Default is \option{red}.}

\item{useLargest}{Optional. If \code{TRUE} then the largest value will be used, i.e. its non-exceedence probability willl be > 0. Default is \code{FALSE}.}
}
\value{
Returns \pkg{ggplot2} object of Korcak plot.
}
\description{
Korcak plot of 1 data set of water areas
}
\examples{
\dontrun{
p <- KorcakPlot1(alldata$areas)}
}
