% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorcakPlot2.R
\name{KorcakPlot2}
\alias{KorcakPlot2}
\title{Korkcak plot of 2 data sets}
\usage{
KorcakPlot2(
  data1,
  name1,
  data2,
  name2,
  minsize = 0,
  xbreaks = c(10, 1000, 1e+05),
  ybreaks = c(0.005, 0.001, 0.01, 0.1, 0.2, 0.4, 0.6, 0.8, 1),
  addLine = "linear",
  useLargest = FALSE
)
}
\arguments{
\item{data1}{Required. Vector holding first data set.}

\item{name1}{Required. Name of first data set.}

\item{data2}{Required. Vector holding second data set.}

\item{name2}{Required. Name of second data set.}

\item{minsize}{Optional. Minimum value to be plotted.}

\item{xbreaks}{Optional. Location of breaks in x-axis. Default is \code{c(10,1000,100000)}}

\item{ybreaks}{Optional. Location of breaks in y-axis. Default is \code{c(0.005,0.001,0.01,0.1,0.2,0.4,0.6,0.8,1)}.}

\item{addLine}{Optional. Default is \option{none}. Other options are \option{linear}, \option{quadratic}, and \option{gpd}.}

\item{useLargest}{Optional. If \code{TRUE} then the largest value will be used, i.e. its non-exceedence probability willl be > 0. Default is \code{FALSE}.}
}
\value{
Returns a \pkg{ggplot2} object of korcak plot.
}
\description{
Korkcak plot of 2 data sets
}
\examples{
\dontrun{
p <- KorcakPlot2(data1, 'water', data2, 'bog')}
}
