% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorcaklPlot2_100m.R
\name{KorcakPlot2_100m}
\alias{KorcakPlot2_100m}
\title{Compares and plots 2 data sets of water patches}
\usage{
KorcakPlot2_100m(data1, name1, data2, name2)
}
\arguments{
\item{data1}{Required. Vector of first set of water areas.}

\item{name1}{Required. String of name of first data set.}

\item{data2}{Required. Vector of second set of water areas.}

\item{name2}{Required. String of name of second data set.}
}
\value{
Returns \pkg{ggplot2} object containing plot.
}
\description{
Only plots patches bigger than 100m x 100m, i.e. 10,000 m2.
}
\examples{
\dontrun{
p <- KorcakPlot2_100m(data1, 'set1', data2, 'set2')}
}
