% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largestWaterPatchVolume.R
\name{largestWaterPatchVolume}
\alias{largestWaterPatchVolume}
\title{Finds volumes of largest water patches}
\usage{
largestWaterPatchVolume(waterFile, threshold = 0.001)
}
\arguments{
\item{waterFile}{Required. An ArcGIS .asc file containing water depths, as produced by \code{WDPM}.}

\item{threshold}{Optional. Threshold for delineating water. Default is 0.001 m.}
}
\value{
Returns a vector consisting of the \code{area} (m\eqn{^2}{^2}), and the \code{volume} (m\eqn{^3}{^3}) for the largest patch.
}
\description{
Finds volumes of largest water patches
}
\examples{
\dontrun{
vols <- waterPatchVolume("5m_300_0_0_0_d.asc", threshold = 0.001)
}
}
