% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop1AreaVolumePlot.R
\name{loop1AreaVolumePlot}
\alias{loop1AreaVolumePlot}
\title{Plots loop1 water area-volume hysteresis}
\usage{
loop1AreaVolumePlot(orderedVA, outfile = "")
}
\arguments{
\item{orderedVA}{Required. Data frame containing WDPM run total water volumes and areas, as returned by \code{volumeAreaSequence}, for the initial loop (complete filling, followed by complete emptying).}

\item{outfile}{Optional. File to hold hysteresis calculated using \code{hysteresis}. If file name is not given, then the hysteresis is output to the screen.}
}
\value{
If successful, returns ggplot2 object of area-volume hysteresis. If unsuccessful, returns \code{FALSE}.
}
\description{
Plots loop1 water area-volume hysteresis
}
\examples{
\dontrun{
p <- loop1AreaVolumePlot}
}
