% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop2ContribFracPlot.R
\name{loop2ContribFracPlot}
\alias{loop2ContribFracPlot}
\title{Plots loop2 contributing fraction hysteresis}
\usage{
loop2ContribFracPlot(orderedAreaVol, extend = FALSE, normalize = TRUE,
  outfile = "", N = 1)
}
\arguments{
\item{orderedAreaVol}{Required. A set of ordered areas and volumes as produced by \code{volumeAreaSequence}}

\item{extend}{Optional. Should curves extend to origin? Default is \code{FALSE}.}

\item{normalize}{Optional. Should areas and volumes be divided by their maximum values? Default is \code{TRUE}.}

\item{outfile}{Optional. File name for output of line segments. If not specified (the default), the data are not ouput.}

\item{N}{Optional. Plot arrows every Nth point. Default is 1.}
}
\value{
Returns a ggplot2 object of the area vs volume. The line segments are written to a file (if specified).
}
\description{
Plots loop2 contributing fraction hysteresis
}
\examples{
\dontrun{
p <- loop2contribFracPlot(areavol)}
}
