% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureBasins.R
\name{measureBasins}
\alias{measureBasins}
\title{Measures areas of basins, as found by Fortran program}
\usage{
measureBasins(infile)
}
\arguments{
\item{infile}{Required. ArcGIS .asc file containing numbered slough basins.}
}
\value{
Returns a data frame containing the slough basin ID, and the areas (m2)
}
\description{
Measures areas of basins, as found by Fortran program
}
\examples{
\dontrun{
areas <- measureBasins('300mm_basins.asc')}
}
