% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureWaterPatches.R
\name{measureWaterPatches}
\alias{measureWaterPatches}
\title{Measures water patches in ArcGIS .asc files}
\usage{
measureWaterPatches(infile, threshold = 0.001)
}
\arguments{
\item{infile}{Required. ArcGIS .asc file of water depths, as output by WDPM.}

\item{threshold}{Optional. Threshold (in m) to determine the existence of water. The default value is 0.001 m, i.e. 1 mm.}
}
\value{
Returns a data frame containing the statistics for each numbered patch
}
\description{
finds the statitics describing water patches in an ArcGIS .asc file.
}
\examples{
\dontrun{
patches <- measureWaterPatches('10_0_0_0_d.asc')}
}
