% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mePlot.R
\name{mePlot}
\alias{mePlot}
\title{Creates mean excess function plot}
\usage{
mePlot(data, cut = 5)
}
\arguments{
\item{data}{Required. Vector to plot.}

\item{cut}{Optional. The number of maxima to exclude. The default value is 5}
}
\value{
Returns \pkg{ggplot2} object.
}
\description{
Creates mean excess function plot
}
\examples{
\dontrun{
p <- mePlot(x)}

}
\references{
From Cirillo, P. (2013), Are your data really Pareto distributed?, Phys. A Stat. Mech. its Appl., 392(23), 5947-5962, doi:10.1016/j.physa.2013.07.061.
}
