% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMareaWidth.R
\name{PCMareaWidth}
\alias{PCMareaWidth}
\title{Calculates width function for graphs of PCM ponds weighted by area}
\usage{
PCMareaWidth(graph, areas)
}
\arguments{
\item{graph}{Required. Graph (from \pkg{igraph}) of PCM ponds. Can be created using \code{graphPCMPonds}.}

\item{areas}{Required. Vector containing pond areas.}
}
\value{
Returns a \code{hist} object of weighted pond counts vs. distance from outlet.
}
\description{
Calculates width function for graphs of PCM ponds weighted by area
}
\examples{
\dontrun{
h <- PCMareaWidth(PCMgraph, areas)
plot(h)}
}
