% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMwidth.R
\name{PCMwidth}
\alias{PCMwidth}
\title{Calculates width function for graphs of PCM ponds}
\usage{
PCMwidth(graph)
}
\arguments{
\item{graph}{Required. Graph (from \pkg{igraph}) of PCM ponds. Can be created using \code{graphPCMPonds}.}
}
\value{
Returns a \code{hist} object of pond counts vs. distance from outlet.
}
\description{
Calculates width function for graphs of PCM ponds
}
\examples{
\dontrun{
h <- PCMwidth(PCMgraph)
plot(h)}
}
