% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contrib_frac_vs_depth.R
\name{plot_contrib_frac_vs_depth}
\alias{plot_contrib_frac_vs_depth}
\title{Plots rising limb contributing fraction vs storage}
\usage{
plot_contrib_frac_vs_depth(orderedAreaVol)
}
\arguments{
\item{orderedAreaVol}{Required. A set of ordered areas and volumes as produced by \code{volumeAreaSequence}}
}
\value{
Returns a list contaning ggplot of the contributing fraction vs depth fo water and the data frame.
}
\description{
Plots rising limb contributing fraction vs storage
}
\examples{
\dontrun{
p <- loop1contribFracPlot(areavol)}
}
