% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCMcontribLoop1.R
\name{plotPCMcontribLoop1}
\alias{plotPCMcontribLoop1}
\title{Plots Loop1 contributing fraction for PCM output}
\usage{
plotPCMcontribLoop1(PCMROPfile, extend = TRUE, normalize = TRUE,
  outfile = "", N = 1)
}
\arguments{
\item{PCMROPfile}{Required. File containing PCM output.}

\item{extend}{Optional. Should curves extend to origin? Default is \code{TRUE}.}

\item{normalize}{Optional. Should areas and volumes be divided by their maximum values? Default is \code{TRUE}.}

\item{outfile}{Optional. File name for output of line segments. If not specified (the default), the data are not ouput.}

\item{N}{Optional. Plot arrows every Nth point. Default is 1.}
}
\value{
Returns a ggplot2 object of the area vs volume. The line segments are written to a file (if specified).
}
\description{
Plots Loop1 contributing fraction for PCM output
}
\examples{
\dontrun{
p <- plotPCMcontribLoop1('AreaVolumeROP_Loop1.txt')}
}
