% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRisingLimbROF.R
\name{plotRisingLimbROF}
\alias{plotRisingLimbROF}
\title{Outputs file of rising limb runoff fraction and volume.}
\usage{
plotRisingLimbROF(sortedAreaVols = "", extend = FALSE, outFile = "")
}
\arguments{
\item{sortedAreaVols}{Required. Set of areas and volumes sorted from smallest to largest.}

\item{extend}{Optional. Logical. Should plots be extended to (0,0) and (1,1)? Default is \code{FALSE}.}

\item{outFile}{Required. CSV file for output.}
}
\value{
Returns rising limb contributing fraction plot.
}
\description{
Outputs file of rising limb runoff fraction and volume.
}
\examples{
\dontrun{p <- plotRisingLimbROF(df, "RisingLimb.csv")
}
}
