% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_roads.R
\name{remove_roads}
\alias{remove_roads}
\title{Removes roads from a DEM}
\usage{
remove_roads(DEM_file, road_shapefile, basin_shapefile)
}
\arguments{
\item{DEM_file}{Required. ArcGIS .asc file.}

\item{road_shapefile}{Required. Shapefile of roads buffered to road widths.}

\item{basin_shapefile}{Required. Shpefile of basin boundary.}
}
\value{
If successful, returns \code{TRUE}. Writes several files, including 1)
a .asc file of the DEM with missing values in place of the roads, 2) SAGA files (.sdat, .mgrd)
contining the filled DEM, and 3) a .asc file of the filled DEM clipped to the basin
boundary.
}
\description{
Removes roads from a DEM
}
\note{
This function uses the \pkg{raster} function \code{mask} to delete the
DEM elements covered by the buffered roads, and \pkg{RSAGA} function \code{rsaga.close.gaps} to fill in
the deleted DEM cells.
}
\examples{
\dontrun{remove_roads("basin5.asc", "basin5_roads.shp", "basin5_boundary.shp")}
}
