% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_length.R
\name{road_length}
\alias{road_length}
\title{Length of roads in a basin}
\usage{
road_length(road_shapefile, basin_shapefile)
}
\arguments{
\item{road_shapefile}{Required. Shapefile containing roads.}

\item{basin_shapefile}{Required. Shapefile of basin boundary.}
}
\value{
Returns a vector of the total length of roads (km), basin area (km2) and
density of roads (km/km2).
}
\description{
Length of roads in a basin
}
\examples{
\dontrun{road_length("basin5roads.shp", "basin5.shp")}
}
