% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_area_volume_patchcount.R
\name{total_area_volume_patchcount}
\alias{total_area_volume_patchcount}
\title{Finds total water volumes, areas and patch counts in WDPM output files}
\usage{
total_area_volume_patchcount(
  location = ".",
  threshold = 0.001,
  filespec = "*.asc",
  outfile = ""
)
}
\arguments{
\item{location}{Optional. Directory containing the WDPM output files. Default is the current working directory.}

\item{threshold}{Optional. Threshold (in m) to determine the existence of water. The default value is 0.001 m, i.e. 1 mm.}

\item{filespec}{Optional. File specification of WDPM output in current working directory. Default is \option{*.asc}.}

\item{outfile}{Optional. Name of file to export values. Values are comma-delimited. If not specified, values are not written.}
}
\value{
Returns basin area, water volume, water area for each .asc file. Optionally writes values to \option{outfile}.
}
\description{
The total area and volume are used to find hysteresis and to determine the contributing (conencted) fraction.
}
\examples{
\dontrun{
setwd('./WDPM/output')
areasVolumes <- totalAreaVolume(filespec='*_d.asc', 'drainedAreasVolumes.csv')}
}
\author{
Kevin Shook
}
