% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapint.R
\name{trapint}
\alias{trapint}
\title{Integrates function by trapeziodal rule}
\usage{
trapint(curve)
}
\arguments{
\item{curve}{Required. A data frame of x and y values. Points must be sorted by x value.}
}
\value{
Returns the area under the curve.
}
\description{
Integrates function by trapeziodal rule
}
\examples{
\dontrun{
area <- trapint(curve)}
}
\author{
Kevin Shook
}
