% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumeAreaSequence.R
\name{volumeAreaSequence}
\alias{volumeAreaSequence}
\title{Orders values created by totalAreaVolume}
\usage{
volumeAreaSequence(areasVolumes)
}
\arguments{
\item{areasVolumes}{Required. A data frame containing WDPM total water areas and volumes.}
}
\value{
If successful, returns a data frame of the water areas and volumes, sorted in order of Add1, Subtract1, Add2, Subtract2. If unsuccssful, returns \code{FALSE}.
}
\description{
The areas and volumes are sorted in order of addition and removal of water. These value can then be used to calculate and plot the hysteresis between water volume and area, and between water volume and contributing (connected) area.
}
\examples{
\dontrun{
areavol <- volumeAreaSequence(areasVolumes)}
}
\seealso{
\code{\link{totalAreaVolume}} \code{\link{loop1ContribFracPlot}}
}
\author{
Kevin Shook
}
