% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterPatchElevation.R
\name{waterPatchElevation}
\alias{waterPatchElevation}
\title{Calculates mean elevations of patches of water}
\usage{
waterPatchElevation(waterFile, threshold = 0.001, DEMfile,
  quiet = TRUE, return = "raster")
}
\arguments{
\item{waterFile}{Required. ArcGIS .asc file of water from WDPM.}

\item{threshold}{Optional. Threshold (in m) to discriminate water. Defaul is 0.001 m.}

\item{DEMfile}{Required. ArcGIS .asc file of DEM elevations.}

\item{quiet}{Optional. If \code{TRUE}, messages are suppressed.}

\item{return}{Optional. If \option{raster}, then returns a raster of the patch elevations.
If \option{patch_vals}, returns a data frame of patch properies. Otherwise, returns both.}
}
\value{
Returns patch mean elevations.
}
\description{
Calculates mean elevations of patches of water
}
\examples{
\dontrun{waterFile = "300_0_0_0_d.asc", DEMfile="basin5.asc")
}
}
