% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterPatchVolume.R
\name{waterPatchVolume}
\alias{waterPatchVolume}
\title{Finds volumes of water patches}
\usage{
waterPatchVolume(patches, waterFile, quiet = TRUE)
}
\arguments{
\item{patches}{Required. \pkg{SDMTools} object containing numbered water patches, as returned by \code{getWaterPatches}.}

\item{waterFile}{Required. An ArcGIS .asc file containing water depths, as produced by \code{WDPM}.}

\item{quiet}{Optional. If \code{TRUE}, the default, there is no extra output. If \code{FALSE}, the patches are listed as they are measured.}
}
\value{
Returns a data frame consisting of the \code{ID}, \code{area} (m\eqn{^2}{^2}), and the \code{volume} (m\eqn{^3}{^3}) for each patch.
}
\description{
Finds volumes of water patches
}
\examples{
\dontrun{
vols <- waterPatchVolume(patches, '5m_300_0_0_0_d.asc')
}
}
