% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zengaPlot.R
\name{zengaPlot}
\alias{zengaPlot}
\title{Create Zenga plot used to examine if data truly fit a Pareto distribution.}
\usage{
zengaPlot(data)
}
\arguments{
\item{data}{Required. Vector of values to be plotted}
}
\value{
Returns \pkg{ggplot2} object.
}
\description{
Create Zenga plot used to examine if data truly fit a Pareto distribution.
}
\examples{
\dontrun{
p <- zengaPlot(x)}
}
\references{
From Cirillo, P. (2013), Are your data really Pareto distributed?, Phys. A Stat. Mech. its Appl., 392(23), 5947-5962, doi:10.1016/j.physa.2013.07.061.
}
